:white_check_mark: CAKE-5021 | Update tests and comment out ios
