+++ package.json
@@ -1,75 +1,78 @@
 {
-  "name": "@wikia/tracking-opt-in",
-  "version": "2.0.0-pre1",
-  "main": "dist/tracking-opt-in.min.js",
-  "license": "Apache-2.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Wikia/tracking-opt-in"
-  },
-  "devDependencies": {
-    "autoprefixer": "^8.4.1",
-    "babel-cli": "^6.26.0",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-istanbul": "^4.1.6",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-preact": "^1.1.0",
-    "browserstack-local": "^1.3.1",
-    "chai": "^4.1.2",
-    "consent-string": "^1.2.4",
-    "cross-env": "^5.1.4",
-    "css-loader": "^0.28.11",
-    "es6-promise": "^4.2.4",
-    "handlebars": "^4.0.11",
-    "handlebars-loader": "^1.7.0",
-    "html-webpack-plugin": "^3.2.0",
-    "js-cookie": "^2.2.0",
-    "jsdom": "^11.10.0",
-    "karma": "^2.0.2",
-    "karma-allure-reporter": "^1.4.4",
-    "karma-coverage": "^1.1.2",
-    "karma-jsdom-launcher": "^6.1.3",
-    "karma-junit-reporter": "^1.2.0",
-    "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.5",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^3.0.0",
-    "mocha": "^5.1.1",
-    "node-sass": "^4.9.0",
-    "postcss-loader": "^2.1.4",
-    "preact": "^8.2.9",
-    "request": "^2.86.0",
-    "rimraf": "^2.6.2",
-    "sass-loader": "^7.0.1",
-    "sinon": "^5.0.7",
-    "style-loader": "^0.21.0",
-    "wdio-allure-reporter": "^0.6.1",
-    "wdio-concise-reporter": "^0.1.2",
-    "wdio-junit-reporter": "^0.4.3",
-    "wdio-mocha-framework": "^0.5.13",
-    "webdriverio": "^4.12.0",
-    "webpack": "^4.6.0",
-    "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-cli": "^2.1.2",
-    "webpack-serve": "^0.3.2"
-  },
-  "scripts": {
-    "nohup:dev": "yarn clean && nohup yarn start:dev &",
-    "nohup:prod": "yarn clean && nohup yarn start:prod &",
-    "dev": "NODE_ENV=development cross-env webpack-serve ./webpack.config.js",
-    "start:dev": "yarn clean && yarn dev",
-    "start:prod": "NODE_ENV=production cross-env webpack-serve ./webpack.config.js",
-    "build": "yarn build:webpack",
-    "build:webpack": "cross-env NODE_ENV=production webpack",
-    "build:babel": "babel src --out-dir dist/src --ignore index-dev.js",
-    "clean": "rimraf dist",
-    "prepublishOnly": "yarn build",
-    "test": "rimraf reports/karma && karma start",
-    "test:watch": "rimraf reports/karma && karma start --auto-watch --no-single-run",
-    "test:selenium": "rimraf reports/webdriver && wdio webdriver.conf.js",
-    "validate-jenkins": "sh release/scripts/validate-jenkins-files.sh"
-  }
+    "name": "@wikia/tracking-opt-in",
+    "version": "2.0.0-pre1",
+    "main": "dist/tracking-opt-in.min.js",
+    "license": "Apache-2.0",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/Wikia/tracking-opt-in"
+    },
+    "devDependencies": {
+        "@wdio/mocha-framework": "^5.11.0",
+        "@wdio/spec-reporter": "^5.11.7",
+        "@wdio/sync": "^5.11.13",
+        "autoprefixer": "^8.4.1",
+        "babel-cli": "^6.26.0",
+        "babel-loader": "^7.1.4",
+        "babel-plugin-istanbul": "^4.1.6",
+        "babel-plugin-transform-class-properties": "^6.24.1",
+        "babel-plugin-transform-object-assign": "^6.22.0",
+        "babel-plugin-transform-object-rest-spread": "^6.26.0",
+        "babel-preset-env": "^1.6.1",
+        "babel-preset-preact": "^1.1.0",
+        "browserstack-local": "^1.3.1",
+        "chai": "^4.1.2",
+        "chromedriver": "^76.0.0",
+        "consent-string": "^1.2.4",
+        "cross-env": "^5.1.4",
+        "css-loader": "^0.28.11",
+        "es6-promise": "^4.2.4",
+        "handlebars": "^4.0.11",
+        "handlebars-loader": "^1.7.0",
+        "html-webpack-plugin": "^3.2.0",
+        "js-cookie": "^2.2.0",
+        "jsdom": "^11.10.0",
+        "karma": "^4.2.0",
+        "karma-allure-reporter": "^1.4.5",
+        "karma-coverage": "^1.1.2",
+        "karma-jsdom-launcher": "^7.1.1",
+        "karma-junit-reporter": "^1.2.0",
+        "karma-mocha": "^1.3.0",
+        "karma-mocha-reporter": "^2.2.5",
+        "karma-sourcemap-loader": "^0.3.7",
+        "karma-webpack": "^4.0.2",
+        "mocha": "^6.2.0",
+        "node-sass": "^4.9.0",
+        "postcss-loader": "^2.1.4",
+        "preact": "^8.2.9",
+        "request": "^2.86.0",
+        "rimraf": "^2.6.2",
+        "sass-loader": "^7.0.1",
+        "sinon": "^5.0.7",
+        "style-loader": "^0.21.0",
+        "webpack": "^4.6.0",
+        "webpack-bundle-analyzer": "^2.13.1",
+        "webpack-cli": "^2.1.2",
+        "webpack-serve": "^0.3.2"
+    },
+    "scripts": {
+        "dev": "NODE_ENV=development cross-env webpack-serve ./webpack.config.js",
+        "start:dev": "yarn clean && yarn dev",
+        "start:prod": "NODE_ENV=production cross-env webpack-serve ./webpack.config.js",
+        "build": "yarn build:webpack",
+        "build:webpack": "cross-env NODE_ENV=production webpack",
+        "build:babel": "babel src --out-dir dist/src --ignore index-dev.js",
+        "clean": "rimraf dist",
+        "prepublishOnly": "yarn build",
+        "test": "rimraf reports/karma && karma start",
+        "test:watch": "rimraf reports/karma && karma start --auto-watch --no-single-run",
+        "test:selenium": "rimraf reports/webdriver && wdio wdio.conf.js",
+        "validate-jenkins": "sh release/scripts/validate-jenkins-files.sh"
+    },
+    "dependencies": {
+        "@wdio/browserstack-service": "^5.11.8",
+        "@wdio/cli": "^5.11.13",
+        "@wdio/local-runner": "^5.11.13",
+        "wdio": "^3.0.3"
+    }
 }
