+++ wdio.conf.js
@@ -1,0 +1,436 @@
+const browserstack = require('browserstack-local');
+const request = require('request');
+
+const user = process.env.BROWSERSTACK_USERNAME;
+const url = process.env.TEST_URL || 'http://localhost:3000';
+const projectName = process.env.PROJECT_NAME || 'tracking-opt-in';
+const key = process.env.BROWSERSTACK_KEY;
+const useTunnel = !!process.env.USE_TUNNEL;
+const build = process.env.BUILD_ID || 'local-test';
+const maxInstances = process.env.MAX_INSTANCES || 3;
+
+const failedTests = [];
+
+const windows10Device = {
+    os: 'Windows',
+    os_version: '10',
+};
+const macOsDevice = {
+    os: 'OS X',
+    os_version: 'High Sierra',
+};
+const android4_4Device = {
+    os: 'Android',
+    device: 'Google Nexus 5',
+    os_version: '4.4',
+    realMobile: true,
+};
+const android5Device = {
+    os: 'Android',
+    device: 'Samsung Galaxy S6',
+    os_version: '5.0',
+    realMobile: true,
+};
+const android6Device = {
+    os: 'Android',
+    device: 'Samsung Galaxy S7',
+    os_version: '6.0',
+    realMobile: true,
+};
+const android7Device = {
+    os: 'Android',
+    device: 'Samsung Galaxy S8',
+    os_version: '7.0',
+    realMobile: true,
+};
+const android8Device = {
+    os: 'Android',
+    device: 'Google Pixel',
+    os_version: '8.0',
+    realMobile: true,
+};
+const ios10_3Device = {
+    os: 'iOS',
+    device: 'iPhone 7',
+    os_version: '10.3',
+    realMobile: true,
+};
+const ios11Device = {
+    os: 'iOS',
+    device: 'iPhone 8',
+    os_version: '11.0',
+    realMobile: true,
+};
+const ios11_2Device = {
+    os: 'iOS',
+    device: 'iPhone SE',
+    os_version: '11.2',
+    realMobile: true,
+};
+const commonCapabilities = {
+    build,
+    project: projectName,
+    'browserstack.local': useTunnel,
+    'browserstack.debug': true,
+    'browserstack.console': 'verbose',
+    'browserstack.user': user,
+    'browserstack.key': key,
+};
+
+const capabilities = [
+    // {
+    //     ...ios10_3Device,
+    //     ...commonCapabilities,
+    //     browser: 'safari',
+    // },
+    // {
+    //     ...ios11Device,
+    //     ...commonCapabilities,
+    //     browser: 'safari',
+    // },
+    // {
+    //     ...ios11_2Device,
+    //     ...commonCapabilities,
+    //     browser: 'safari',
+    // },
+    {
+        ...android4_4Device,
+        ...commonCapabilities,
+        browser: 'chrome',
+    },
+    {
+        ...android5Device,
+        ...commonCapabilities,
+        browser: 'chrome',
+    },
+    {
+        ...android6Device,
+        ...commonCapabilities,
+        browser: 'chrome',
+    },
+    {
+        ...android7Device,
+        ...commonCapabilities,
+        browser: 'chrome',
+    },
+    {
+        ...android8Device,
+        ...commonCapabilities,
+        browser: 'chrome',
+    },
+    {
+        ...windows10Device,
+        ...commonCapabilities,
+        browser: 'chrome',
+    },
+    {
+        ...windows10Device,
+        ...commonCapabilities,
+        browser: 'firefox',
+    },
+    {
+        ...windows10Device,
+        ...commonCapabilities,
+        browser: 'edge',
+    },
+    {
+        ...macOsDevice,
+        ...commonCapabilities,
+        browser: 'firefox',
+    },
+    {
+        ...macOsDevice,
+        ...commonCapabilities,
+        browser: 'safari',
+    },
+    {
+        ...macOsDevice,
+        ...commonCapabilities,
+        browser: 'chrome',
+    }
+];
+
+exports.config = {
+    //
+    // ====================
+    // Runner Configuration
+    // ====================
+    //
+    // WebdriverIO allows it to run your tests in arbitrary locations (e.g. locally or
+    // on a remote machine).
+    runner: 'local',
+    //
+    // Override default path ('/wd/hub') for chromedriver service.
+    // path: '/',
+    // port: 9515,
+    //
+    // =================
+    // Service Providers
+    // =================
+    // WebdriverIO supports Sauce Labs, Browserstack, and Testing Bot (other cloud providers
+    // should work too though). These services define specific user and key (or access key)
+    // values you need to put in here in order to connect to these services.
+    //
+    user: user,
+    key: key,
+    //
+    // If you run your tests on SauceLabs you can specify the region you want to run your tests
+    // in via the `region` property. Available short handles for regions are `us` (default) and `eu`.
+    // These regions are used for the Sauce Labs VM cloud and the Sauce Labs Real Device Cloud.
+    // If you don't provide the region it will default for the `us`
+
+    //
+    // ==================
+    // Specify Test Files
+    // ==================
+    // Define which test specs should run. The pattern is relative to the directory
+    // from which `wdio` was called. Notice that, if you are calling `wdio` from an
+    // NPM script (see https://docs.npmjs.com/cli/run-script) then the current working
+    // directory is where your package.json resides, so `wdio` will be called from there.
+    //
+    specs: [
+        `./selenium/*.js`,
+    ],
+    // Patterns to exclude.
+    exclude: [
+        // 'path/to/excluded/files'
+    ],
+    //
+    // ============
+    // Capabilities
+    // ============
+    // Define your capabilities here. WebdriverIO can run multiple capabilities at the same
+    // time. Depending on the number of capabilities, WebdriverIO launches several test
+    // sessions. Within your capabilities you can overwrite the spec and exclude options in
+    // order to group specific specs to a specific capability.
+    //
+    // First, you can define how many instances should be started at the same time. Let's
+    // say you have 3 different capabilities (Chrome, Firefox, and Safari) and you have
+    // set maxInstances to 1; wdio will spawn 3 processes. Therefore, if you have 10 spec
+    // files and you set maxInstances to 10, all spec files will get tested at the same time
+    // and 30 processes will get spawned. The property handles how many capabilities
+    // from the same test should run tests.
+    //
+    maxInstances: 10,
+    //
+    // If you have trouble getting all important capabilities together, check out the
+    // Sauce Labs platform configurator - a great tool to configure your capabilities:
+    // https://docs.saucelabs.com/reference/platforms-configurator
+    //
+    capabilities: capabilities,
+    browserstackLocal: true,
+    //
+    // ===================
+    // Test Configurations
+    // ===================
+    // Define all options that are relevant for the WebdriverIO instance here
+    //
+    // Level of logging verbosity: trace | debug | info | warn | error | silent
+    logLevel: 'info',
+    //
+    // Set specific log levels per logger
+    // loggers:
+    // - webdriver, webdriverio
+    // - @wdio/applitools-service, @wdio/browserstack-service, @wdio/devtools-service, @wdio/sauce-service
+    // - @wdio/mocha-framework, @wdio/jasmine-framework
+    // - @wdio/local-runner, @wdio/lambda-runner
+    // - @wdio/sumologic-reporter
+    // - @wdio/cli, @wdio/config, @wdio/sync, @wdio/utils
+    // Level of logging verbosity: trace | debug | info | warn | error | silent
+    // logLevels: {
+    //     webdriver: 'info',
+    //     '@wdio/applitools-service': 'info'
+    // },
+    //
+    // If you only want to run your tests until a specific amount of tests have failed use
+    // bail (default is 0 - don't bail, run all tests).
+    bail: 0,
+    //
+    // Set a base URL in order to shorten url command calls. If your `url` parameter starts
+    // with `/`, the base url gets prepended, not including the path portion of your baseUrl.
+    // If your `url` parameter starts without a scheme or `/` (like `some/path`), the base url
+    // gets prepended directly.
+    baseUrl: url,
+    //
+    // Default timeout for all waitFor* commands.
+    waitforTimeout: 10000,
+    //
+    // Default timeout in milliseconds for request
+    // if Selenium Grid doesn't send response
+    connectionRetryTimeout: 90000,
+    //
+    // Default request retries count
+    connectionRetryCount: 3,
+    //
+    // Test runner services
+    // Services take over a specific job you don't want to take care of. They enhance
+    // your test setup with almost no effort. Unlike plugins, they don't add new
+    // commands. Instead, they hook themselves up into the test process.
+    services: ['browserstack'],
+
+    // Framework you want to run your specs with.
+    // The following are supported: Mocha, Jasmine, and Cucumber
+    // see also: https://webdriver.io/docs/frameworks.html
+    //
+    // Make sure you have the wdio adapter package for the specific framework installed
+    // before running any tests.
+    framework: 'mocha',
+    //
+    // The number of times to retry the entire specfile when it fails as a whole
+    // specFileRetries: 1,
+    //
+    // Test reporter for stdout.
+    // The only one supported by default is 'dot'
+    // see also: https://webdriver.io/docs/dot-reporter.html
+    reporters: ['spec'],
+
+    //
+    // Options to be passed to Mocha.
+    // See the full list at http://mochajs.org/
+    mochaOpts: {
+        ui: 'bdd',
+        timeout: 60000
+    },
+    //
+    // =====
+    // Hooks
+    // =====
+    // WebdriverIO provides several hooks you can use to interfere with the test process in order to enhance
+    // it and to build services around it. You can either apply a single function or an array of
+    // methods to it. If one of them returns with a promise, WebdriverIO will wait until that promise got
+    // resolved to continue.
+    /**
+     * Gets executed once before all workers get launched.
+     * @param {Object} config wdio configuration object
+     * @param {Array.<Object>} capabilities list of capabilities details
+     */
+    onPrepare: function (config, capabilities) {
+        console.log("Connecting local");
+        return new Promise(function(resolve, reject){
+            exports.bs_local = new browserstack.Local();
+            exports.bs_local.start({'key': exports.config.key, 'forceLocal': true }, function(error) {
+                if (error) return reject(error);
+                console.log('Connected. Now testing...');
+
+                resolve();
+            });
+        });
+    },
+    /**
+     * Gets executed just before initialising the webdriver session and test framework. It allows you
+     * to manipulate configurations depending on the capability or spec.
+     * @param {Object} config wdio configuration object
+     * @param {Array.<Object>} capabilities list of capabilities details
+     * @param {Array.<String>} specs List of spec file paths that are to be run
+     */
+    // beforeSession: function (config, capabilities, specs) {
+    // },
+    /**
+     * Gets executed before test execution begins. At this point you can access to all global
+     * variables like `browser`. It is the perfect place to define custom commands.
+     * @param {Array.<Object>} capabilities list of capabilities details
+     * @param {Array.<String>} specs List of spec file paths that are to be run
+     */
+    // before: function (capabilities, specs) {
+    // },
+    /**
+     * Runs before a WebdriverIO command gets executed.
+     * @param {String} commandName hook command name
+     * @param {Array} args arguments that command would receive
+     */
+    // beforeCommand: function (commandName, args) {
+    // },
+    /**
+     * Hook that gets executed before the suite starts
+     * @param {Object} suite suite details
+     */
+    // beforeSuite: function (suite) {
+    // },
+    /**
+     * Function to be executed before a test (in Mocha/Jasmine) or a step (in Cucumber) starts.
+     * @param {Object} test test details
+     */
+    // beforeTest: function (test) {
+    // },
+    /**
+     * Hook that gets executed _before_ a hook within the suite starts (e.g. runs before calling
+     * beforeEach in Mocha)
+     */
+    // beforeHook: function () {
+    // },
+    /**
+     * Hook that gets executed _after_ a hook within the suite starts (e.g. runs after calling
+     * afterEach in Mocha)
+     */
+    // afterHook: function () {
+    // },
+    /**
+     * Function to be executed after a test (in Mocha/Jasmine) or a step (in Cucumber) starts.
+     * @param {Object} test test details
+     */
+    afterTest: function (test) {
+        if (!test.passed) {
+            failedTests.push(test.fullTitle);
+        }
+    },
+    /**
+     * Hook that gets executed after the suite has ended
+     * @param {Object} suite suite details
+     */
+    // afterSuite: function (suite) {
+    // },
+    /**
+     * Runs after a WebdriverIO command gets executed
+     * @param {String} commandName hook command name
+     * @param {Array} args arguments that command would receive
+     * @param {Number} result 0 - command success, 1 - command error
+     * @param {Object} error error object if any
+     */
+    // afterCommand: function (commandName, args, result, error) {
+    // },
+    /**
+     * Gets executed after all tests are done. You still have access to all global variables from
+     * the test.
+     * @param {Number} result 0 - test pass, 1 - test fail
+     * @param {Array.<Object>} capabilities list of capabilities details
+     * @param {Array.<String>} specs List of spec file paths that ran
+     */
+    after: function (result, capabilities, specs) {
+        if (result !== 0) {
+            request({
+                uri: `https://${user}:${key}@api.browserstack.com/automate/sessions/${browser.sessionId}.json`,
+                method: 'PUT',
+                form: {
+                    status: 'error',
+                    reason: failedTests.join(';'),
+                }
+            });
+        }
+    },
+    /**
+     * Gets executed right after terminating the webdriver session.
+     * @param {Object} config wdio configuration object
+     * @param {Array.<Object>} capabilities list of capabilities details
+     * @param {Array.<String>} specs List of spec file paths that ran
+     */
+    // afterSession: function (config, capabilities, specs) {
+    // },
+    /**
+     * Gets executed after all workers got shut down and the process is about to exit. An error
+     * thrown in the onComplete hook will result in the test run failing.
+     * @param {Object} exitCode 0 - success, 1 - fail
+     * @param {Object} config wdio configuration object
+     * @param {Array.<Object>} capabilities list of capabilities details
+     * @param {<Object>} results object containing test results
+     */
+    onComplete: function(exitCode, config, capabilities, results) {
+        exports.bs_local.stop(function() {});
+    },
+    /**
+     * Gets executed when a refresh happens.
+     * @param {String} oldSessionId session ID of the old session
+     * @param {String} newSessionId session ID of the new session
+     */
+    //onReload: function(oldSessionId, newSessionId) {
+    //}
+}
